refactor: Tool to my-tools module
